Add files via upload
This is the MZI draft layout completed as part of SiEPEC - Draft Layout Submission. Note that the appended username cael3 corresponds with "cael_3" username on edX. I could not make the file naming consistent since underscores are used to represent different device characteristics in the op layout names, and my edX username would create a conflict with this naming convention.