Submitting additional file EBeam_VATamma_E.gds
Submitting additional file EBeam_VATamma_E.gds
Please note the only error is related to 'Fiber Array Configuration'. This is by design as I have named 3 of 4 GC's with separate text labels. Not sure how to convince the DRC that this is OK and will work with the automated testing equipment. The idea is based on one of the videos in the class lectures.