Two different design files (Three copies of each)
Each files of EBeam_Mehdi_paperclips_3 and EBeam_Mehdi_paperclips_4 has two more copies (..._2nd and ..._3rd). The input port names for the copied files are different to be able to differentiate the results. However, the measurements should be supposedly very similar. They only will be used to calculate the error bars.