Submitting additional files to be merged - 7 file (_I.gds to _O.gds)
Hello, Submitting additional files to be merged - 7 file (_I.gds to _O.gds)
Note for the TE paperclip I see the following error: "Radius [EBeam_VATamma_]
Not enough space to accommodate the desired bend radius for the waveguide."
During the zoom call on 07 Oct Prof said it was OK to have this error and it can be waived. I waive the 'radius' error for the paperclip spiral.