Ring Resonators_with Verification errors
There are errors on RR5 and RR6. The reason is because I wanted two ring resonators (RR5 and RR6) include three and two rings in series, respectively. But closer than simply possible by connecting the defined ports. So, I was not sure how I could reduce the straight waveguide of each half-ring directional coupler to make the rings in series closer. This is the source of error. Is it still possible to fabricate them?